estoque-api/frontend/requirements.txt
Flask==2.0.1
Flask-Bootstrap4==4.0.2
requests==2.26.0
python-dotenv==0.19.0
Werkzeug==2.0.1
Jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
urllib3==1.26.6
WTForms==2.3.3
Flask-WTF==0.15.1
Bootstrap-Flask==2.0.1
